Support setting global rsync config bind address
Adds the configuration parameter: address to
the rsync::server class.

This configuration parameter can be used to specify
the value of the rsync global config setting: address.

This commit also converts the global headers from a hard
coded file source to a template to support the
parameterization of the address configuration setting.