Make rsync::get idempotent
This commit makes the rsync::get define idenpotent. It does this by adding
a dry-run rsync call as on onlyif params that ensures that rsync will
only create an event if the file currently needs to be updated.

This change has been added to ensure that the rsync::get define does not
constantly spew events when its used.