Clean up Debian patches
The original pull request for Debian support was a little old,
so some of the coding conventions have been cleaned up a little bit:
case sentivity, selectors instead of case statements, etc.

Also included a requirement on puppetlabs/stdlib in order to
use the function str2bool, which is handy.