Fixes for server header
Allow address to be undef so that rsyncd can listen on all interfaces
useful for IPv4/IPv6 dual stack boxes where you want the daemon to
listen on both simultaneously. By setting address to :: the server
was only listening on IPv6 but setting it to 0.0.0.0 it was only
listening on IPv4. Leaving it out of the config makes the daemon
listen on all interfaces at startup.